requirements.txt
amazon.ion==0.5.0
boto3==1.9.237
botocore==1.12.237
ionhash==1.1.0
pytest==4.6.3
pytest-cov==2.7.1